# ironbridge-rust/docs/assets/openapi.yaml
openapi: 3.0.3
info:
  title: IronBridge API
  description: |
    REST API for IronBridge (Chat Session Manager) - a unified platform for harvesting, 
    managing, and analyzing AI chat sessions across multiple providers and tools.

    ## Authentication

    Currently the API supports local access without authentication. Enterprise features
    include SSO/SAML authentication and API key-based access.

    ## Rate Limiting

    Rate limiting is optional and configurable per-installation.

    ## WebSocket

    Real-time updates are available via WebSocket at `ws://<host>/ws`. Note that
    `/ws` is mounted at the server root, not under the `/api` base URL below.

    Client frames are tagged JSON: `subscribe`, `unsubscribe`, `ping`,
    `sync_request`, `stream_start`, `stream_cancel`, `stream_input` and
    `agent_command`. Every one of them is answered.

    `sync_request` returns the same delta as `GET /sync/delta?from=N`, as one
    `sync_event` per change in version order. Two cases answer with an error
    instead of an empty result, because sending nothing would read as "you are
    up to date": `history_truncated` when the requested version has aged out of
    the retained event history, and `version_ahead` when the client's version
    exceeds the server's. Both mean the client needs a full snapshot.

    The stream operations and `agent_command` are **not implemented** on this
    server and answer `stream_error` / `error` with code `unsupported`. They are
...

...

...
servers:
...
paths:
...
  /webhooks:
    servers: *rootServer
    get:
      tags: [Webhooks]
      summary: List webhooks
      description: This scope answers bare, without the `{success, data}` envelope.
      operationId: listWebhooks
      responses:
        "200":
          description: Webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    post:
      tags: [Webhooks]
      summary: Register a webhook
      operationId: createWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Registered
...
  /webhooks/{id}:
    servers: *rootServer
    get:
      tags: [Webhooks]
      summary: Get a webhook
      operationId: getWebhook
      parameters:
        - $ref: "#/components/parameters/sessionId"
      responses:
        "200":
          description: Webhook
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Webhooks]
      summary: Update a webhook
      operationId: updateWebhook
      parameters:
        - $ref: "#/components/parameters/sessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Webhooks]
      summary: Delete a webhook
      operationId: deleteWebhook
      parameters:
        - $ref: "#/components/parameters/sessionId"
      responses:
        "200":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /webhooks/{id}/test:
    servers: *rootServer
    post:
      tags: [Webhooks]
      summary: Send a test delivery
      operationId: testWebhook
      parameters:
        - $ref: "#/components/parameters/sessionId"
      responses:
        "200":
          description: Test dispatched
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    sessionId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
      required: [error, message]
...